+++ src/main/java/com/dart/api/application/auth/AuthenticationService.java
@@ -12,10 +12,10 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import com.dart.api.domain.member.entity.Member;
 import com.dart.api.domain.member.repo.MemberRepository;
-import com.dart.dto.auth.TokenReqDto;
-import com.dart.dto.auth.TokenResDto;
-import com.dart.dto.member.request.LoginReqDto;
-import com.dart.dto.member.response.LoginResDto;
+import com.dart.api.dto.auth.TokenReqDto;
+import com.dart.api.dto.auth.TokenResDto;
+import com.dart.api.dto.member.request.LoginReqDto;
+import com.dart.api.dto.member.response.LoginResDto;
 import com.dart.global.error.exception.BadRequestException;
 import com.dart.global.error.exception.NotFoundException;
 import com.dart.global.error.model.ErrorCode;
