+++ src/main/java/com/dart/api/application/gallery/GalleryService.java
@@ -20,8 +20,8 @@ import com.dart.api.domain.gallery.repo.GalleryRepository;
 import com.dart.api.domain.gallery.repo.HashtagRepository;
 import com.dart.api.domain.member.entity.Member;
 import com.dart.api.domain.member.repo.MemberRepository;
-import com.dart.dto.gallery.request.CreateGalleryDto;
-import com.dart.dto.gallery.request.DeleteGalleryDto;
+import com.dart.api.dto.gallery.request.CreateGalleryDto;
+import com.dart.api.dto.gallery.request.DeleteGalleryDto;
 import com.dart.global.common.util.S3Service;
 import com.dart.global.error.exception.BadRequestException;
 import com.dart.global.error.exception.NotFoundException;
