+++ src/main/java/com/dart/api/application/payment/PaymentService.java
@@ -20,9 +20,9 @@ import com.dart.api.domain.member.entity.Member;
 import com.dart.api.domain.member.repo.MemberRepository;
 import com.dart.api.domain.payment.entity.Payment;
 import com.dart.api.domain.payment.repo.PaymentRepository;
-import com.dart.dto.payment.request.PaymentCreateDto;
-import com.dart.dto.payment.response.PaymentApproveDto;
-import com.dart.dto.payment.response.PaymentReadyDto;
+import com.dart.api.dto.payment.request.PaymentCreateDto;
+import com.dart.api.dto.payment.response.PaymentApproveDto;
+import com.dart.api.dto.payment.response.PaymentReadyDto;
 import com.dart.global.config.PaymentProperties;
 import com.dart.global.error.exception.NotFoundException;
 import com.dart.global.error.model.ErrorCode;
