+++ src/main/java/com/dart/api/domain/review/entity/Review.java
@@ -15,8 +15,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import com.dart.api.domain.gallery.entity.Gallery;
 import com.dart.api.domain.member.entity.Member;
-import com.dart.dto.review.request.ReviewCreateDto;
-import com.dart.dto.review.response.ReviewReadDto;
+import com.dart.api.dto.review.request.ReviewCreateDto;
+import com.dart.api.dto.review.response.ReviewReadDto;
 import com.dart.global.common.entity.BaseTimeEntity;
 
 @Entity
