+++ src/main/java/com/dart/api/dto/auth/TokenReqDto.java
@@ -1,0 +1,6 @@
+package com.dart.api.dto.auth;
+
+public record TokenReqDto(
+	String expiredToken
+) {
+}
