+++ src/main/java/com/dart/api/dto/auth/TokenResDto.java
@@ -1,0 +1,9 @@
+package com.dart.api.dto.auth;
+
+import lombok.Builder;
+
+@Builder
+public record TokenResDto(
+	String accessToken
+) {
+}
