+++ src/main/java/com/dart/api/dto/gallery/request/ImageInfoDto.java
@@ -1,0 +1,7 @@
+package com.dart.api.dto.gallery.request;
+
+public record ImageInfoDto(
+	String imageTitle,
+	String description
+) {
+}
