+++ src/main/java/com/dart/api/dto/member/response/LoginResDto.java
@@ -1,0 +1,6 @@
+package com.dart.api.dto.member.response;
+
+public record LoginResDto(
+	String accessToken
+) {
+}
