+++ src/main/java/com/dart/api/dto/member/response/MemberProfileResDto.java
@@ -1,0 +1,13 @@
+package com.dart.api.dto.member.response;
+
+public record MemberProfileResDto(
+	String email,
+	String nickname,
+	String profileImage,
+
+	String bank,
+	String account,
+	String introduce
+){
+
+}
