+++ src/main/java/com/dart/api/dto/payment/response/PaymentApproveDto.java
@@ -1,0 +1,9 @@
+package com.dart.api.dto.payment.response;
+
+import java.time.LocalDateTime;
+
+public record PaymentApproveDto(
+	String item_code,
+	LocalDateTime approved_at
+) {
+}
