+++ src/main/java/com/dart/api/dto/payment/response/PaymentReadyDto.java
@@ -1,0 +1,7 @@
+package com.dart.api.dto.payment.response;
+
+public record PaymentReadyDto(
+	String tid,
+	String next_redirect_pc_url
+) {
+}
