+++ src/main/java/com/dart/api/dto/review/response/PageInfo.java
@@ -1,0 +1,7 @@
+package com.dart.api.dto.review.response;
+
+public record PageInfo(
+	int pageIndex,
+	boolean isDone
+) {
+}
