+++ src/main/java/com/dart/api/dto/review/response/PageResponse.java
@@ -1,0 +1,9 @@
+package com.dart.api.dto.review.response;
+
+import java.util.List;
+
+public record PageResponse(
+	List<ReviewReadDto> reviews,
+	PageInfo pageInfo
+) {
+}
