+++ src/main/java/com/dart/api/dto/review/response/ReviewReadDto.java
@@ -1,0 +1,16 @@
+package com.dart.api.dto.review.response;
+
+import java.time.LocalDateTime;
+
+import lombok.Builder;
+
+@Builder
+public record ReviewReadDto(
+	Long reviewId,
+	LocalDateTime createAt,
+	String content,
+	int score,
+	String nickname,
+	String profileImage
+) {
+}
