--- src/main/java/com/dart/dto/auth/TokenReqDto.java
@@ -1,6 +1,0 @@
-package com.dart.dto.auth;
-
-public record TokenReqDto(
-	String expiredToken
-) {
-}
