--- src/main/java/com/dart/dto/auth/TokenResDto.java
@@ -1,9 +1,0 @@
-package com.dart.dto.auth;
-
-import lombok.Builder;
-
-@Builder
-public record TokenResDto(
-	String accessToken
-) {
-}
