--- src/main/java/com/dart/dto/gallery/request/ImageInfoDto.java
@@ -1,7 +1,0 @@
-package com.dart.dto.gallery.request;
-
-public record ImageInfoDto(
-	String imageTitle,
-	String description
-) {
-}
