--- src/main/java/com/dart/dto/member/request/NicknameDuplicationCheckDto.java
@@ -1,9 +1,0 @@
-package com.dart.dto.member.request;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record NicknameDuplicationCheckDto(
-	@NotBlank(message = "[❎ ERROR] 닉네임을 입력해주세요.")
-	String nickname
-) {
-}
