--- src/main/java/com/dart/dto/member/response/LoginResDto.java
@@ -1,6 +1,0 @@
-package com.dart.dto.member.response;
-
-public record LoginResDto(
-	String accessToken
-) {
-}
