--- src/main/java/com/dart/dto/member/response/MemberProfileResDto.java
@@ -1,13 +1,0 @@
-package com.dart.dto.member.response;
-
-public record MemberProfileResDto(
-	String email,
-	String nickname,
-	String profileImage,
-
-	String bank,
-	String account,
-	String introduce
-){
-
-}
