--- src/main/java/com/dart/dto/payment/response/PaymentApproveDto.java
@@ -1,9 +1,0 @@
-package com.dart.dto.payment.response;
-
-import java.time.LocalDateTime;
-
-public record PaymentApproveDto(
-	String item_code,
-	LocalDateTime approved_at
-) {
-}
