--- src/main/java/com/dart/dto/payment/response/PaymentReadyDto.java
@@ -1,7 +1,0 @@
-package com.dart.dto.payment.response;
-
-public record PaymentReadyDto(
-	String tid,
-	String next_redirect_pc_url
-) {
-}
