--- src/main/java/com/dart/dto/review/response/PageInfo.java
@@ -1,7 +1,0 @@
-package com.dart.dto.review.response;
-
-public record PageInfo(
-	int pageIndex,
-	boolean isDone
-) {
-}
