--- src/main/java/com/dart/dto/review/response/PageResponse.java
@@ -1,9 +1,0 @@
-package com.dart.dto.review.response;
-
-import java.util.List;
-
-public record PageResponse(
-	List<ReviewReadDto> reviews,
-	PageInfo pageInfo
-) {
-}
